HR_Attendance_using_Face_Detection/requirements.txt
Opencv-cv2
numpy
base64
face_recognition
PIL
tkinter